api/pyproject.toml
[project]
name = "graphlagoon"
version = "0.8.4"
description = "Graph visualization and exploration tool with FastAPI backend"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "AGPL-3.0" }
authors = [
    { name = "Bruno Messias", email = "devmessias@gmail.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Visualization",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "sqlglot>=26.0.0",
    "gsql2rsql",
    "asyncpg>=0.31.0",
]

[project.optional-dependencies]
# PostgreSQL support (optional - database can be disabled)
postgres = [
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
]
# Databricks Lakebase support
lakebase = [
    "databricks-sdk>=0.38.0",
]
# All optional dependencies
all = [
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "databricks-sdk>=0.38.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "debugpy>=1.8.0",
    "ruff>=0.4.0",
]

[project.scripts]
graphlagoon = "graphlagoon.cli:main"

[project.urls]
Homepage = "https://github.com/graphlagoon/graphlagoon"
Documentation = "https://github.com/graphlagoon/graphlagoon#readme"
Repository = "https://github.com/graphlagoon/graphlagoon"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["graphlagoon"]

[tool.uv.sources]
gsql2rsql = { path = "../../cyper2dsql/python", editable = true }
